Revert last change, move everything to a different package, added end portal fix.
