+++ core/src/main/java/org/geysermc/geyser/translator/collision/fixes/EndPortalCollision.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019-2022 GeyserMC. http://geysermc.org
+ * Copyright (c) 2025 GeyserMC. http://geysermc.org
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,7 +23,7 @@
  * @link https://github.com/GeyserMC/Geyser
  */
 
-package org.geysermc.geyser.translator.collision;
+package org.geysermc.geyser.translator.collision.fixes;
 
 import lombok.EqualsAndHashCode;
 import org.geysermc.geyser.level.block.property.Properties;
@@ -32,40 +32,30 @@ import org.geysermc.geyser.level.physics.BoundingBox;
 import org.geysermc.geyser.level.physics.CollisionManager;
 import org.geysermc.geyser.level.physics.Direction;
 import org.geysermc.geyser.session.GeyserSession;
+import org.geysermc.geyser.translator.collision.BlockCollision;
+import org.geysermc.geyser.translator.collision.CollisionRemapper;
 
 @EqualsAndHashCode(callSuper = true)
-@CollisionRemapper(regex = "_door$", usesParams = true, passDefaultBoxes = true)
-public class DoorCollision extends BlockCollision {
-    /**
-     * 1 = north
-     * 2 = east
-     * 3 = south
-     * 4 = west
-     */
-    private int facing;
+@CollisionRemapper(regex = "^end_portal_frame$", passDefaultBoxes = true)
+public class EndPortalCollision extends BlockCollision {
+    private final boolean eye;
 
-    public DoorCollision(BlockState state, BoundingBox[] defaultBoxes) {
-        super(defaultBoxes);
-        facing = switch (state.getValue(Properties.HORIZONTAL_FACING)) {
-            case NORTH -> 1;
-            case EAST -> 2;
-            case SOUTH -> 3;
-            case WEST -> 4;
-            default -> throw new IllegalStateException();
-        };
+    public EndPortalCollision(BlockState state, BoundingBox[] boxes) {
+        super(boxes);
 
-        // If the door is open it changes direction
-        if (state.getValue(Properties.OPEN)) {
-            facing = facing % 2 + 1;
-        }
+        this.eye = state.getValue(Properties.EYE);
     }
 
     @Override
     public void correctPosition(GeyserSession session, int x, int y, int z, BoundingBox playerCollision) {
         super.correctPosition(session, x, y, z, playerCollision);
-        final double maxPushDistance = 0.005 + CollisionManager.COLLISION_TOLERANCE * 1.01F;
+        if (!this.eye) {
+            return;
+        }
 
-        // Check for door bug (doors are 0.1875 blocks thick on Java but 0.1825 blocks thick on Bedrock)
+        final double maxPushDistance = 0.1875F + CollisionManager.COLLISION_TOLERANCE * 1.01F;
+
+        // Check for end portal frame bug (BE don't have the eye collision when the end portal frame contain an eye unlike JE)
         for (BoundingBox boundingBox : this.boundingBoxes) {
             if (!boundingBox.checkIntersection(x, y, z, playerCollision)) {
                 continue;
@@ -74,12 +64,7 @@ public class DoorCollision extends BlockCollision {
             boundingBox = boundingBox.clone();
             boundingBox.translate(x, y, z);
 
-            switch (this.facing) {
-                case 1 -> boundingBox.pushOutOfBoundingBox(playerCollision, Direction.NORTH, maxPushDistance);
-                case 2 -> boundingBox.pushOutOfBoundingBox(playerCollision, Direction.EAST, maxPushDistance);
-                case 3 -> boundingBox.pushOutOfBoundingBox(playerCollision, Direction.SOUTH, maxPushDistance);
-                case 4 -> boundingBox.pushOutOfBoundingBox(playerCollision, Direction.WEST, maxPushDistance);
-            }
+            boundingBox.pushOutOfBoundingBox(playerCollision, Direction.UP, maxPushDistance);
         }
     }
 }
